.history/requirements_20250912164815.txt
# Vietnamese Medical Chatbot - Python Dependencies
# Generated for notebooks 3 & 4 submission

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
pymysql>=1.1.0
alembic>=1.13.0

# Machine Learning & NLP
torch>=2.1.0
transformers>=4.35.0
tokenizers>=0.15.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# API Integration
requests>=2.31.0
python-dotenv>=1.0.0

# Export Functionality
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Development & Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.26.0

# Optional: Enhanced functionality
python-multipart>=0.0.6  # For file uploads
python-jose[cryptography]>=3.3.0  # For JWT tokens (if needed)
passlib[bcrypt]>=1.7.4  # For password hashing (if needed)

# Vietnamese Text Processing (Optional)
underthesea>=6.7.0  # Vietnamese NLP toolkit